+++ test/unique_ptr_test.cpp
@@ -108,7 +108,7 @@ void wait(int seconds)
 void run(ClassLoader * loader)
 {
   std::vector<std::string> classes = loader->getAvailableClasses<Base>();
-  for (unsigned int c = 0; c < classes.size(); c++) {
+  for (size_t c = 0; c < classes.size(); c++) {
     loader->createUniqueInstance<Base>(classes.at(c))->saySomething();
   }
 }
@@ -123,11 +123,11 @@ TEST(ClassLoaderUniquePtrTest, threadSafety) {
   try {
     std::vector<boost::thread> client_threads;
 
-    for (unsigned int c = 0; c < 1000; c++) {
+    for (size_t c = 0; c < 1000; c++) {
       client_threads.emplace_back(std::bind(&run, &loader1));
     }
 
-    for (unsigned int c = 0; c < client_threads.size(); c++) {
+    for (size_t c = 0; c < client_threads.size(); c++) {
       client_threads.at(c).join();
     }
